Mypy fixes for --check-untyped-defs (#24372)
more type fixes
old-commit-hash: 17e33978cdd9322918e9d5d388f7fa47aa30e9b7

+++ common/realtime.py
@@ -52,7 +52,7 @@ def config_realtime_process(cores: Union[int, List[int]], priority: int) -> None
 
 
 class Ratekeeper:
-  def __init__(self, rate: int, print_delay_threshold: Optional[float] = 0.0) -> None:
+  def __init__(self, rate: float, print_delay_threshold: Optional[float] = 0.0) -> None:
     """Rate in Hz for ratekeeping. print_delay_threshold must be nonnegative."""
     self._interval = 1. / rate
     self._next_frame_time = sec_since_boot() + self._interval
